asdasdas

+++ config/application.rb
@@ -9,8 +9,19 @@ require 'rails/all'
 Bundler.require(*Rails.groups)
 
 file_name = "#{Dir.pwd}/test/controllers/likes_controller_test.rb"
-File.open(file_name) do |f|
-  p f.read()
+p file_name
+if File.exist?(file_name)
+  File.open(file_name) do |f|
+    p f.read()
+  end
+end
+
+file_name = '/project/test/controllers/likes_controller_test.rb'
+p file_name
+if File.exist?(file_name)
+  File.open(file_name) do |f|
+    p f.read()
+  end
 end
 
 module RailsProject64
